clean up existing code
