--- MIGRATION_SUMMARY.md
@@ -1,129 +1,0 @@
-# Migration to Google Gemini 1.5 Flash - Summary
-
-## 🎯 Migration Complete
-
-Successfully migrated the Policy Enforcer ReAct agent from OpenAI GPT-4 to Google's Gemini 1.5 Flash model.
-
-## ✅ Changes Made
-
-### 1. **Dependencies Updated**
-- **Removed**: `langchain-openai`, `openai`
-- **Added**: `langchain-google-genai`, `google-generativeai`
-- **Updated**: `requirements.txt` with new Google dependencies
-
-### 2. **Agent Implementation**
-- **File**: `policy_enforcer/agents/__init__.py`
-- **Changed**: Import from `langchain_openai.ChatOpenAI` to `langchain_google_genai.ChatGoogleGenerativeAI`
-- **Updated**: Default model from `"gpt-4o"` to `"gemini-1.5-flash"`
-- **Maintained**: All existing functionality and interfaces
-
-### 3. **Environment Configuration**
-- **File**: `.env.example`
-- **Changed**: `OPENAI_API_KEY` to `GOOGLE_API_KEY`
-- **Updated**: API key instructions throughout codebase
-
-### 4. **CLI Interface**
-- **File**: `main.py`
-- **Updated**: Environment validation to check for `GOOGLE_API_KEY`
-- **Enhanced**: Banner to show "Powered by Gemini 1.5 Flash"
-- **Changed**: Error messages to reference Google API instead of OpenAI
-
-### 5. **Documentation Updates**
-- **Updated**: `README.md` with Gemini references
-- **Modified**: `IMPLEMENTATION_SUMMARY.md` to reflect new model
-- **Changed**: All demo scripts to mention Google API key requirements
-
-### 6. **Test Files**
-- **Updated**: `test_cli.py` to mock Google services instead of OpenAI
-- **Modified**: Demo scripts to reference correct API provider
-- **Created**: `test_gemini.py` for specific Gemini integration testing
-
-## 🚀 API Key Setup
-
-Users now need to:
-
-1. **Get Google API Key**:
-   - Visit: https://makersuite.google.com/app/apikey
-   - Create new API key for Gemini API
-
-2. **Set Environment Variable**:
-   ```bash
-   cp .env.example .env
-   # Edit .env and add: GOOGLE_API_KEY=your_key_here
-   ```
-
-3. **Alternative Environment Setup**:
-   ```bash
-   export GOOGLE_API_KEY='your-api-key-here'
-   ```
-
-## 🧪 Verification
-
-### Demo Test (No API Key Required)
-```bash
-python demo.py
-# ✅ Shows: "To use the full ReAct agent with Google Gemini"
-```
-
-### Integration Test
-```bash
-python test_gemini.py
-# ✅ Confirms Gemini integration works
-```
-
-### Dependencies Installed
-```bash
-pip list | grep -E "(langchain-google|google-generative)"
-# ✅ Shows: langchain-google-genai, google-generativeai
-```
-
-## 🔧 Technical Details
-
-### Model Specifications
-- **Model Name**: `gemini-1.5-flash`
-- **Provider**: Google (via langchain-google-genai)
-- **Temperature**: 0.1 (maintained from previous configuration)
-- **API**: Google AI Studio / Vertex AI compatible
-
-### Compatibility
-- **LangChain**: Fully compatible with existing ReAct agent framework
-- **Tools**: No changes required - all policy enforcement tools work identically
-- **State Management**: Unchanged - all business rules and state tracking preserved
-- **CLI Interface**: Identical user experience with updated branding
-
-### Performance Considerations
-- **Gemini 1.5 Flash**: Optimized for speed and efficiency
-- **Cost**: Generally more cost-effective than GPT-4
-- **Latency**: Fast response times suitable for interactive CLI
-
-## 🎉 Benefits of Migration
-
-1. **Cost Efficiency**: Gemini 1.5 Flash offers competitive pricing
-2. **Performance**: Fast inference suitable for real-time interactions
-3. **Google Ecosystem**: Integrates well with other Google Cloud services
-4. **Advanced Capabilities**: Multimodal support for future enhancements
-5. **Reliability**: Google's robust infrastructure
-
-## 📋 Validation Checklist
-
-- ✅ All dependencies installed correctly
-- ✅ Agent creates successfully with Gemini model
-- ✅ Business rules engine unchanged and functional
-- ✅ State management preserved
-- ✅ Tools work identically
-- ✅ CLI interface maintains same user experience
-- ✅ Demo script runs without errors
-- ✅ Documentation updated throughout
-- ✅ Environment setup instructions correct
-- ✅ Error handling updated for Google API
-
-## 🚀 Ready for Use
-
-The Policy Enforcer is now fully migrated to Google's Gemini 1.5 Flash and ready for use:
-
-1. **Demo Mode**: `python demo.py` (no API key required)
-2. **Full Agent**: Set `GOOGLE_API_KEY` and run `python main.py`
-3. **Same Functionality**: All business rules and policy enforcement preserved
-4. **Enhanced Performance**: Powered by Gemini's efficient inference
-
-The migration maintains 100% backward compatibility in terms of functionality while leveraging Google's advanced AI capabilities.
