--- REFACTORING_SUMMARY.md
@@ -1,148 +1,0 @@
-# Code Refactoring Summary - Prompt Comparison Feature
-
-## Problem Statement
-The original prompt comparison implementation had significant code duplication across multiple files:
-- `save_prompts.py` duplicated prompt generation logic from the agent module
-- `export_prompts.py` and `save_prompts.py` performed similar functions
-- `demo_prompt_comparison.py` was overly complex with unnecessary features
-
-## Refactoring Solution
-
-### 1. Centralized Prompt Generation
-**Created:** `policy_enforcer/prompt_utils.py`
-- Single source of truth for all prompt generation logic
-- Eliminates duplication between agent and utility scripts
-- Provides clean API for prompt comparison
-
-**Key Functions:**
-```python
-generate_prompt_template(include_rules: bool) -> str    # Core generation
-compare_prompts() -> Dict[str, Any]                     # Compare both versions  
-quick_export(include_rules: Optional[bool]) -> List[str] # Export to files
-save_prompt_to_file(include_rules: bool, file: str)     # Save with metadata
-```
-
-### 2. Updated Agent Module
-**Modified:** `policy_enforcer/agents/__init__.py`
-- Now imports and uses `generate_prompt_template()` from `prompt_utils`
-- Removed 60+ lines of duplicated prompt generation code
-- Maintains identical functionality
-
-**Before:**
-```python
-# 60+ lines of duplicated prompt generation logic
-if self.include_rules_in_prompt:
-    rules_section = f"""..."""
-    # ... extensive duplication
-```
-
-**After:**
-```python
-from ..prompt_utils import generate_prompt_template
-prompt_template = generate_prompt_template(self.include_rules_in_prompt)
-```
-
-### 3. Unified Export Utility
-**Created:** `prompt_export.py` (replaces 2 files)
-- Single, simple utility replacing `save_prompts.py` and `export_prompts.py`
-- Clean command-line interface
-- Generates both individual files and comparison reports
-
-**Usage:**
-```bash
-python prompt_export.py              # Export both + comparison
-python prompt_export.py --with-rules # Export only with-rules version
-python prompt_export.py --no-rules   # Export only learning mode
-python prompt_export.py --compare-only # Only comparison report
-```
-
-### 4. Simplified Demo
-**Simplified:** `demo_prompt_comparison.py`
-- Reduced from 180+ lines to 30 lines
-- Focused on core functionality demonstration
-- Removed complex workflow demos
-
-**Before:** Complex multi-section demo with argument parsing tests
-**After:** Simple comparison stats and file export demo
-
-## Files Removed
-- ❌ `save_prompts.py` (314 lines) - functionality moved to `prompt_utils.py`
-- ❌ `export_prompts.py` (90 lines) - functionality moved to `prompt_export.py`
-
-## Files Created/Modified
-
-### New Files
-- ✅ `policy_enforcer/prompt_utils.py` (150 lines) - Centralized utilities
-- ✅ `prompt_export.py` (80 lines) - Unified export tool
-- ✅ `PROMPT_COMPARISON_README.md` - Simple documentation
-
-### Modified Files  
-- ✅ `policy_enforcer/agents/__init__.py` - Uses centralized prompt generation
-- ✅ `demo_prompt_comparison.py` - Simplified from 180→30 lines
-- ✅ `README.md` - Updated features list
-
-## Benefits Achieved
-
-### Code Quality
-- **-424 lines** total code reduction
-- **Zero duplication** of prompt generation logic  
-- **Single source of truth** for prompt templates
-- **Simplified maintenance** - changes in one place
-
-### User Experience
-- **Simpler CLI** with clear options
-- **Faster execution** - no redundant processing
-- **Better documentation** with focused README
-- **Consistent behavior** across all tools
-
-### Research Workflow
-- **Quick exports** for prompt comparison
-- **Easy command-line switching** between modes
-- **Clean comparison reports** with statistics
-- **Reliable functionality** backed by existing tests
-
-## Validation
-
-### ✅ All Tests Pass
-- 101 unit tests continue to pass
-- No breaking changes to existing functionality
-- Agent behavior identical to before refactoring
-
-### ✅ Feature Parity
-- Main application works with `--rules` and `--no-rules` flags
-- Prompt export generates identical output
-- Demo shows same core functionality
-- All ablation study features preserved
-
-### ✅ Code Cleanliness
-- No code duplication detected
-- Clear separation of concerns
-- Intuitive file organization
-- Comprehensive documentation
-
-## Usage Summary
-
-### For Researchers
-```bash
-# Generate prompt comparison
-python prompt_export.py
-
-# Run experiments
-python main.py --rules          # Explicit rules mode
-python main.py --no-rules       # Learning mode
-
-# Quick demo
-python demo_prompt_comparison.py
-```
-
-### For Developers
-```python
-# Import centralized utilities
-from policy_enforcer.prompt_utils import generate_prompt_template, compare_prompts
-
-# Generate prompts
-prompt_with_rules = generate_prompt_template(True)
-prompt_learning = generate_prompt_template(False)
-```
-
-The refactoring successfully eliminated code duplication while maintaining all functionality and improving the developer and researcher experience.
