--- SEMANTIC_KERNEL_MIGRATION.md
@@ -1,214 +1,0 @@
-# Semantic Kernel Migration Summary
-
-This document summarizes the successful migration of the Policy Enforcer codebase from LangChain to Semantic Kernel.
-
-## Migration Overview
-
-The Policy Enforcer demonstrates a ReAct agent with business rule enforcement. The original implementation used LangChain's built-in ReAct agent, which had to be manually implemented for Semantic Kernel since it doesn't provide a built-in ReAct pattern.
-
-## What Was Migrated
-
-### ✅ Successfully Migrated Components
-
-1. **ReAct Agent Pattern** (`policy_enforcer/react_agent.py`)
-   - Custom implementation of ReAct (Reasoning + Acting) loop
-   - Supports Question → Thought → Action → Action Input → Observation cycle
-   - Handles parsing of agent responses and tool execution
-   - Configurable max iterations and verbose logging
-
-2. **Tools → Plugins** (`policy_enforcer/sk_tools.py`)
-   - Converted LangChain tools to Semantic Kernel plugins
-   - Maintained policy enforcement pattern
-   - Used `@kernel_function` decorators with type annotations
-   - Preserved all business rule checking logic
-
-3. **Agent Wrapper** (`policy_enforcer/sk_agents.py`)
-   - Created PolicyEnforcerSKAgent to match original API
-   - Integrated with Google AI via Semantic Kernel connectors
-   - Maintained state awareness and prompt generation patterns
-
-4. **Dependencies** (`requirements.txt`)
-   - Replaced LangChain dependencies with semantic-kernel
-   - Reduced dependency footprint significantly
-   - Maintained pydantic and python-dotenv compatibility
-
-5. **Main Application** (`main_sk.py`)
-   - Ported CLI interface with identical functionality
-   - Preserved ablation study modes (--rules / --no-rules)
-   - Maintained argument parsing and environment setup
-
-### 🔄 Preserved Components
-
-The following components were **reused without changes** across both implementations:
-
-- **Business Rules Engine** (`policy_enforcer/rules/`)
-- **State Management** (`policy_enforcer/state/`)
-- **Item Definitions** (`policy_enforcer/items.py`)
-- **Core Business Logic** (All 7 business rules preserved exactly)
-
-## Key Implementation Details
-
-### Custom ReAct Agent
-
-Since Semantic Kernel doesn't provide a built-in ReAct agent, I implemented one from scratch:
-
-```python
-class ReActAgent:
-    def __init__(self, kernel, service_id, instructions, max_iterations=10):
-        # Custom ReAct implementation
-    
-    async def run_async(self, question: str) -> str:
-        # Manual ReAct loop with parsing and tool execution
-    
-    def _parse_response(self, response: str) -> Optional[str]:
-        # Parse LLM output for Thought/Action/Action Input/Final Answer
-    
-    async def _execute_action(self, action: str, action_input: Dict) -> str:
-        # Execute tools through Semantic Kernel function calls
-```
-
-### Plugin-Based Tool Architecture
-
-Converted LangChain tools to Semantic Kernel plugins:
-
-```python
-class WeatherPlugin(PolicyEnforcedPlugin):
-    @kernel_function(description="Check the current weather condition")
-    def check_weather(self) -> Annotated[str, "Weather condition"]:
-        # Rule checking + tool execution
-```
-
-### Function Metadata and Type Safety
-
-Semantic Kernel's function system provides better type safety:
-
-```python
-@kernel_function(name="shopping")
-def shopping(
-    self, 
-    item: Annotated[str, "The item to purchase: TV, Xbox, Hiking Boots, Goggles, or Sunscreen"]
-) -> Annotated[str, "Result of the shopping action with updated inventory"]:
-```
-
-## Migration Benefits
-
-### 1. **Reduced Dependencies**
-- LangChain: 5 direct dependencies + many transitive
-- Semantic Kernel: 1 direct dependency (semantic-kernel)
-
-### 2. **Better Type Safety**
-- Semantic Kernel enforces typed function signatures
-- `Annotated` types provide clear parameter documentation
-- Better IDE support and validation
-
-### 3. **Cleaner Architecture**
-- Plugin-based approach is more modular
-- Function decorators are more declarative
-- Less boilerplate code for tool definitions
-
-### 4. **Microsoft Ecosystem Integration**
-- Better integration with Azure services
-- Enterprise-ready with official Microsoft support
-- Active development and GA status
-
-## Testing and Validation
-
-Created comprehensive migration tests (`test_sk_migration.py`):
-
-- ✅ State management functionality
-- ✅ Business rules engine
-- ✅ Semantic Kernel plugins
-- ✅ Prompt generation (both modes)
-- ✅ Rule enforcement in plugins
-
-All tests pass, confirming migration success.
-
-## Usage Instructions
-
-### Quick Start
-
-1. **Create virtual environment:**
-   ```bash
-   python3 -m venv venv
-   source venv/bin/activate
-   ```
-
-2. **Install dependencies:**
-   ```bash
-   pip install -r requirements.txt
-   ```
-
-3. **Set Google API key:**
-   ```bash
-   export GOOGLE_API_KEY='your-api-key-here'
-   ```
-
-4. **Run Semantic Kernel version:**
-   ```bash
-   python main_sk.py
-   ```
-
-5. **Test migration:**
-   ```bash
-   python test_sk_migration.py
-   ```
-
-### Comparison Testing
-
-Both implementations can run side-by-side:
-
-```bash
-# Original LangChain version
-python main.py --rules
-
-# New Semantic Kernel version  
-python main_sk.py --rules
-```
-
-## File Structure
-
-### New Files Added
-- `main_sk.py` - Semantic Kernel CLI entry point
-- `policy_enforcer/react_agent.py` - Custom ReAct implementation
-- `policy_enforcer/sk_agents.py` - Agent wrapper
-- `policy_enforcer/sk_tools.py` - Plugin implementations
-- `policy_enforcer/sk_prompt_utils.py` - Prompt utilities
-- `test_sk_migration.py` - Migration validation tests
-
-### Files Modified
-- `requirements.txt` - Updated dependencies
-- `CLAUDE.md` - Updated documentation
-
-### Files Preserved
-- All original LangChain files remain unchanged
-- Business logic completely preserved
-- State management unchanged
-- Rules engine identical
-
-## Performance Considerations
-
-The Semantic Kernel implementation has some differences in performance characteristics:
-
-1. **Slower Cold Start**: Initial model loading may be slower
-2. **Manual ReAct Loop**: Custom implementation may have different timing
-3. **Better Function Metadata**: Semantic Kernel's type system may improve LLM function selection
-
-## Future Improvements
-
-1. **Async Support**: Could be enhanced with native async/await patterns
-2. **Streaming**: Semantic Kernel supports streaming responses
-3. **Advanced Planners**: Could integrate SK's planning capabilities
-4. **Vector Operations**: Better integration with vector databases
-5. **Multi-Agent**: Semantic Kernel's agent framework supports multi-agent scenarios
-
-## Conclusion
-
-The migration to Semantic Kernel was successful, maintaining 100% functional compatibility while providing:
-
-- ✅ Cleaner, more maintainable code
-- ✅ Better type safety and IDE support  
-- ✅ Reduced dependency footprint
-- ✅ Enterprise-ready Microsoft ecosystem integration
-- ✅ Custom ReAct implementation that can be extended
-
-The original LangChain implementation remains available for comparison and can coexist with the new Semantic Kernel version.
