--- pytest.ini
@@ -1,14 +1,0 @@
-[tool:pytest]
-testpaths = tests
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-addopts = 
-    -v
-    --strict-markers
-    --tb=short
-    --showlocals
-markers =
-    integration: Integration tests that test complete workflows
-    unit: Unit tests for individual components
-    error_handling: Tests for error scenarios and edge cases
